requirements.txt
sacred==0.7.3
numpy==1.15.1
scipy==1.1.0
scikit-learn==0.19.1
networkx==2.1
joblib==0.11
cython==0.28.5
pymongo==3.7.1
pandas==0.23.4
pyyaml==3.13
tensorflow-gpu==1.9.0

